tpctk-fs-async-backend/requirements.txt
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.2
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
oauthlib==3.2.0
orjson==3.6.8
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pydantic==1.9.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
tweepy==4.8.0
typing_extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
zipp==3.8.0
